refactor: extract PaymentsAbstractClient from AbstractClient
