+++ src/Clients/StoresClient.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Montonio\Clients;
 
-class StoresClient extends AbstractClient
+class StoresClient extends PaymentsAbstractClient
 {
     public function getPaymentMethods(): array
     {
