+++ src/MontonioClient.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Montonio;
 
-use Montonio\Clients\AbstractClient;
 use Montonio\Clients\OrdersClient;
+use Montonio\Clients\PaymentsAbstractClient;
 use Montonio\Clients\PaymentIntentsClient;
 use Montonio\Clients\PaymentLinksClient;
 use Montonio\Clients\PayoutsClient;
@@ -13,7 +13,7 @@ use Montonio\Clients\RefundsClient;
 use Montonio\Clients\SessionsClient;
 use Montonio\Clients\StoresClient;
 
-class MontonioClient extends AbstractClient
+class MontonioClient extends PaymentsAbstractClient
 {
     public const ENVIRONMENT_SANDBOX = 'sandbox';
     public const ENVIRONMENT_LIVE = 'live';
